+++ cli/commands/seed.go
@@ -3,6 +3,8 @@ package commands
 import (
 	"fmt"
 
+	"github.com/hasura/graphql-engine/cli/seed"
+
 	"github.com/hasura/graphql-engine/cli"
 	"github.com/hasura/graphql-engine/cli/internal/hasura"
 	"github.com/hasura/graphql-engine/cli/internal/metadatautil"
@@ -75,4 +77,13 @@ func NewSeedCmd(ec *cli.ExecutionContext) *cobra.Command {
 	util.BindPFlag(v, "certificate_authority", f.Lookup("certificate-authority"))
 
 	return seedCmd
+}
+
+func getSeedDriver(configVersion cli.ConfigVersion) (driver *seed.Driver) {
+	if configVersion >= cli.V3 {
+		driver = seed.NewDriver(ec.APIClient.V2Query.Bulk, ec.APIClient.PGDump)
+	} else {
+		driver = seed.NewDriver(ec.APIClient.V1Query.Bulk, ec.APIClient.PGDump)
+	}
+	return driver
 }