+++ cli/migrate/database/driver.go
@@ -117,8 +117,6 @@ type Driver interface {
 
 	SchemaDriver
 
-	SeedDriver
-
 	SettingsDriver
 
 	Query(data interface{}) error